moving Switcher to its own file
